refactor: validation.ts -> schema.ts로 명칭 변경
