+++ src/features/login-form/index.ts
@@ -1,2 +1,2 @@
-export { type LoginInput, LoginSchema } from "@/features/login-form/model/validation";
+export { type LoginInput, LoginSchema } from "@/features/login-form/model/schema";
 export { LoginForm } from "@/features/login-form/ui/LoginForm";
